Add: three-jet observable data structures
Use Mathematica associations (hash maps) to store all of the information about
the event. Legs and dipoles hold information about the geometry of the event in
a convenient form and are observable independent. Further I use a map to hold
all of the information about the observable for a given event, but this must be
provided for each observable that one would like to resum.