Add: Observable handling
Add the complete structure for the D-parameter and all of the functions that are
needed for its computation. Include a skeleton structure for the C-parameter
too.

Start writing interface for the initialisation of a generic observable.
Initialisation must cover set up of pointers to appropriate functions and to
load any grids that are needed for the multiple emission functions.