Fix: initialisation of observables
Had some leaky abstraction, initialisation routines for specific observables
were visible to the user. Change the structure of the observable specific
initialisation routines to avoid this, so that in the generic initialisation
routine we can just do a Needs[] at the top with all the observables and just
select the ones that we want to use.