Add: convenience interface for events
Build events that contain the leg and dipole configurations. Also include
information on xq and xqb. This idea could be extended to configurations with
different numbers of jets. Idea is going to be to use this event interface
rather than either xq and xqb or legs and dipoles in the more high level
routines.