Refactor: resummation interface
Change radiator so that we work in terms of xmuR instead of muR and Q.
This simplifies the interface somewhat.

Change resummation so that it accepts alphas, xmuR and logXV as inputs rather
than computing them inside. These calculations will be moved into a driver
program. This way the resummation routines contain only the assembly of the full
resummation.