Refactor: expansion and resummation
Checks of (resummation - expansion) work for LL and NLL with the variation of
xmuR. Great refactor moves around terms so that the LL radiator also has the
terms from the expansion of as(Qab) and so on. Due to dipole nature of expansion
change the soft radiator expansion so that Gnm functions return a list of
dipoles. This way I can convolve this with the appropriate scales.